+++ src/backend/Kys.Application/Resources/Queries/GetSharedResources/GetSharedResourcesQueryHandler.cs
@@ -13,6 +13,7 @@ public sealed class GetSharedResourcesQueryHandler(IResourceRepository repositor
             r.Id,
             r.Name,
             r.Description,
+            r.ResourceTypeId,
             r.ResourceType.Name,
             r.ResourceType.Code,
             r.EnvironmentScope,